requirements.txt
# Core dependencies
streamlit==1.29.0
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2
pydantic==2.5.2
pydantic-settings==2.1.0
qrcode==7.4.2
Pillow==10.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Additional utilities
numpy==1.24.3
python-dotenv==1.0.0
uuid==1.30
pytz==2023.3
xlsxwriter==3.1.9

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
